Renamed aofs to assembler
