--- aofs/__init__.py
@@ -1,2 +1,0 @@
-from aofs.parser import parse_line # Function to parse a line of assembly
-from aofs.hexfile import to_hexfile # Function to parse an assembly file to hex file
