+++ assembler/parser.py
@@ -1,13 +1,7 @@
-from aofs.token_set import token_set
+from assembler.token_set import token_set
 
-# Parse a token into integer code (i.e. index of token)
-def parse_token(tokens : list, token_type : str) -> int:
-
-    # reads from the first token
-    if tokens == []:
-        raise Exception(f"Expected a {token_type} token.")
-    token = tokens[0]
-    tokens.pop(0)
+# Parse a mnemonic into integer code (essentially just an index-of)
+def parse_mnemonic(token : str, token_type : str) -> int:
 
     if token_type not in token_set:
         raise Exception(f"Token type {token_type} does not exist.")
@@ -20,13 +14,7 @@ def parse_token(tokens : list, token_type : str) -> int:
         raise Exception(f"Unrecognized {token_type} '{token}'.")
 
 # Parse an immediate into integer with the specified bitsize
-def parse_imm(tokens : list, bit_size : int) -> int:
-
-    # reads from the first token
-    if tokens == []:
-        raise Exception(f"Expected a {bit_size}-bit immediate value.")
-    token = tokens[0]
-    tokens.pop(0)
+def parse_immediate(token : str, bit_size : int) -> int:
 
     result = 0
     try:
@@ -46,42 +34,43 @@ def parse_imm(tokens : list, bit_size : int) -> int:
 
 # Parses a line of assembly to integer code
 def parse_line(asmline : str) -> int:
+
+    # firstly, handle comma split into header and operands
     asmline = asmline.strip()
     tokens = asmline.split()
 
-    # firstly, get the opcode and rd code
-    opcode = parse_token(tokens, "instruction")
+    # next, parse the header (the opcode and destination register)
+    opcode = parse_mnemonic(tokens[0], "instruction")
+    rd = parse_mnemonic(tokens[1], "register") << 8
 
+    # next, parse the operands
     # parsing operands are finicky; syntax is specific to opcode
-    # nevertheless, it always starts with Destination Register
-    rd = parse_token(tokens, "register")
-    operands = rd << 8
 
     # start with conditionals ("move" and "cadd")
     if opcode in [0, 1]:
-        rn = parse_token(tokens, "register")
-        cnd = parse_token(tokens, "condition")
-        operands += (rn << 4) + cnd
+        rn = parse_mnemonic(tokens[2], "register")
+        cnd = parse_mnemonic(tokens[3], "condition")
+        operands = (rn << 4) + cnd
     # next, do memory instructions ("load" and "store")
     elif opcode in [2, 3]:
-        rn = parse_token(tokens, "register")
-        imm4 = parse_imm(tokens, 4)
-        operands += (rn << 4) + imm4
+        rn = parse_mnemonic(tokens[2], "register")
+        imm4 = parse_immediate(tokens[3], 4)
+        operands = (rn << 4) + imm4
     # insert instruction takes 8-bit immediate
     elif opcode == 4:
-        imm8 = parse_imm(tokens, 8)
-        operands += imm8
+        imm8 = parse_immediate(tokens[2], 8)
+        operands = imm8
     # shift instruction takes a special "shift operation" token type
     elif opcode == 5:
-        shiftop = parse_token(tokens, "shift operation")
-        distance = parse_token(tokens, "register")
-        operands += (shiftop << 4) + distance
+        shiftop = parse_mnemonic(tokens[2], "shift operation")
+        distance = parse_mnemonic(tokens[3], "register")
+        operands = (shiftop << 4) + distance
     # rest ALU instructions take registers
     else:
-        rn = parse_token(tokens, "register")
-        rm = parse_token(tokens, "register")
-        operands += (rn << 4) + rm
+        rn = parse_mnemonic(tokens[2], "register")
+        rm = parse_mnemonic(tokens[3], "register")
+        operands = (rn << 4) + rm
 
-    bincode = (opcode << 12) + operands
+    bincode = (opcode << 12) + (rd << 8) + operands
 
     return bincode
