+++ main.py
@@ -1,9 +1,9 @@
-import sys, aofs
+import sys, assembler
 
 if __name__ == "__main__":
 
     if len(sys.argv) != 2:
         raise Exception("Requires second argument for file path")
     
-    aofs.to_hexfile(sys.argv[1])
+    assembler.to_hexfile(sys.argv[1])
